+++ Rebus.Tests/Profiling/TestProfiler.cs
@@ -4,6 +4,7 @@ using System.Threading.Tasks;
 using NUnit.Framework;
 using Rebus.Messages;
 using Rebus.Pipeline;
+using Rebus.Pipeline.Invokers;
 using Rebus.Profiling;
 using Rebus.Tests.Contracts;
 using Rebus.Transport;